src/qlib_/requirements.txt
pandas
numpy
scikit-learn
mlflow
chronos-forecasting
pyqlib
xgboost
catboost
lightgbm